Upload staged-voice-control 1.1.0.825 [4394]

+++ source/lp/COMPONENT_IFX_LPWWD/source/cy_lpwwd_postwwd.c
@@ -67,11 +67,11 @@
  ******************************************************/
 #if ENABLE_LPWWD_LOGS == 2
 #define cy_lpwwd_postwwd_log_info(format,...)  printf ("[postwwd] "format" \r\n",##__VA_ARGS__);
-#define cy_lpwwd_postwwd_log_err(ret_val,format,...)  printf ("[postwwd] [Err:0x%"PRIx32", %s:%d] "format" \r\n",ret_val,__FUNCTION__,__LINE__,##__VA_ARGS__);
+#define cy_lpwwd_postwwd_log_err(ret_val,format,...)  printf ("[postwwd] [Err:0x%"PRIx32", %s:%d] "format" \r\n",(uint32_t)ret_val,__FUNCTION__,__LINE__,##__VA_ARGS__);
 #define cy_lpwwd_postwwd_log_dbg(format,...)  printf ("[postwwd] "format" \r\n",##__VA_ARGS__);
 #elif ENABLE_LPWWD_LOGS
 #define cy_lpwwd_postwwd_log_info(format,...)  cy_log_msg (CYLF_MIDDLEWARE,CY_LOG_INFO,"[postwwd] "format" \r\n",##__VA_ARGS__);
-#define cy_lpwwd_postwwd_log_err(ret_val,format,...)  cy_log_msg (CYLF_MIDDLEWARE,CY_LOG_INFO,"[postwwd] [Err:0x%"PRIx32", %s:%d] "format" \r\n",ret_val,__FUNCTION__,__LINE__,##__VA_ARGS__);
+#define cy_lpwwd_postwwd_log_err(ret_val,format,...)  cy_log_msg (CYLF_MIDDLEWARE,CY_LOG_INFO,"[postwwd] [Err:0x%"PRIx32", %s:%d] "format" \r\n",(uint32_t)ret_val,__FUNCTION__,__LINE__,##__VA_ARGS__);
 #define cy_lpwwd_postwwd_log_dbg(format,...)  cy_log_msg (CYLF_MIDDLEWARE,CY_LOG_INFO,"[postwwd] "format" \r\n",##__VA_ARGS__);
 #else
 #define cy_lpwwd_postwwd_log_info(format,...)
